+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # Changelog
+
+## Version 0.1
+* initial release
+* utils from old zhinst package copied
+* deviceutils from old zhinst-deviceutils package copied for
+  * SHFQA
+  * SHFSG
+  * SHFQC